* master: (144 commits)
  Applied some minor changes to examples/makefile.
  There is a segmentation fault on load_file("MeshFiles/unittest/io/fv26x46x72.t.3.nc", NULL, "VARIABLE=US"). It turns out that some tags of ScdInterface class (e.g. __BOX_SET) might get deleted (e.g. by Core::clean_up_failed_read) before they are dereferenced later. These dangling tags should be reset to NULL so that they can be recreated upon request.
  fix distcheck failure cubit 14 file added to the EXTRA_DIST variable
  cubit 14 test for global id also, correct the path for tqdcfr test; maybe it should be moved to proper test folder.
  example to load partially a file, based on tagged sets
  Minor changes to structuredmesh and StructuredMeshSimple examples.
  Changes to eliminate requirement for MPI2, from Russ Whitesides (LLNL).
  another format change for cubit 14
  Fixed a problem with entities having the character '#' in their name, which messed up scanning of lines in the embedded ACIS file.
  enhance crystal router example
  add a crystal router example
  Applied some minor changes to examples/makefile and added a few more ignore items.
  Fixed some build errors of point_in_elem_search example (caused by recent merged spatial locator code).
  Fixed compilation errors of structuredmesh and StructuredMeshSimple examples.
  Adding a few more ignore items.
  valgrind find and typo in TupleList for intersection test, valgrind reported a memory leak that memory is cleaned in destructor now in TupleList, wrong size was used in reporting reallocation error
  Adding file to the to-be-removed list after tests.
  Fixing a bunch of warnings, mainly about narrowing conversions.
  Fixing errors turned up by make check.
  Enhanced the HOMME reader so that it can read a connectivity file directly, and updated unit test read_ucd_nc for this new feature. Note, a HOMME connectivity file normally has enough information to create the mesh, though some variables are not available (e.g. lev, time and T).
  ...

Conflicts:
	src/CMakeLists.txt
	src/parallel/CMakeLists.txt
	tools/CMakeLists.txt
	tools/dagmc/CMakeLists.txt
	tools/mbcoupler/CMakeLists.txt

+++ test/h5file/h5portable.cpp
@@ -107,7 +107,7 @@ void create_mesh( const char* filename )
   
   
   for (size_t i = 0; i < NUM_VERT; ++i) {
-    double coords[] = { i % 9, i / 9, Z };
+    double coords[] = { static_cast<double>(i % 9), static_cast<double>(i / 9), static_cast<double>(Z) };
     rval = mb.create_vertex( coords, verts[i] );
     CHECK_ERR(rval);
   }
